fix tests
problem of SpringBoot contexts when mixing JUnit4 & JUnit5
