+++ src/test/java/org/gridsuite/voltageinit/server/service/parameters/VoltageInitParametersTest.java
@@ -4,10 +4,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.voltageinit.server;
+package org.gridsuite.voltageinit.server.service.parameters;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.powsybl.iidm.network.IdentifiableType;
 import com.powsybl.iidm.network.Network;
@@ -24,34 +22,29 @@ import org.assertj.core.api.Condition;
 import org.assertj.core.api.ListAssert;
 import org.gridsuite.voltageinit.server.dto.parameters.FilterEquipments;
 import org.gridsuite.voltageinit.server.dto.parameters.IdentifiableAttributes;
-import org.gridsuite.voltageinit.server.dto.parameters.VoltageInitParametersInfos;
-import org.gridsuite.voltageinit.server.dto.parameters.VoltageLimitInfos;
 import org.gridsuite.voltageinit.server.entities.parameters.FilterEquipmentsEmbeddable;
 import org.gridsuite.voltageinit.server.entities.parameters.VoltageInitParametersEntity;
 import org.gridsuite.voltageinit.server.entities.parameters.VoltageLimitEntity;
-import org.gridsuite.voltageinit.server.repository.parameters.VoltageInitParametersRepository;
 import org.gridsuite.voltageinit.server.service.VoltageInitRunContext;
-import org.gridsuite.voltageinit.server.service.parameters.FilterService;
-import org.gridsuite.voltageinit.server.service.parameters.VoltageInitParametersService;
 import org.gridsuite.voltageinit.server.util.VoltageLimitParameterType;
 import org.gridsuite.voltageinit.utils.TestUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.skyscreamer.jsonassert.Customization;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 import org.skyscreamer.jsonassert.comparator.CustomComparator;
 import org.skyscreamer.jsonassert.comparator.JSONComparator;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureTestEntityManager;
+import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.annotation.DirtiesContext.ClassMode;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -60,26 +53,18 @@ import java.util.UUID;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.condition.NestableCondition.nestable;
 import static org.assertj.core.condition.VerboseCondition.verboseCondition;
-import static org.gridsuite.voltageinit.utils.assertions.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-/**
- * @author Ayoub LABIDI <ayoub.labidi at rte-france.com>
- */
+@ExtendWith({ MockitoExtension.class })
 @SpringBootTest
-@DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
-@AutoConfigureMockMvc
-@Transactional
+@DirtiesContext
+@AutoConfigureTestEntityManager
+@Transactional(propagation = Propagation.REQUIRES_NEW)
 @Slf4j
 class VoltageInitParametersTest {
-    private static final String URI_PARAMETERS_BASE = "/v1/parameters";
-    private static final String URI_PARAMETERS_GET_PUT = URI_PARAMETERS_BASE + "/";
-
     private static final UUID NETWORK_UUID = UUID.fromString("7928181c-7977-4592-ba19-88027e4254e4");
     private static final String VARIANT_ID_1 = "variant_1";
     private static final UUID FILTER_UUID_1 = UUID.fromString("1a3d23a6-7a4c-11ee-b962-0242ac120002");
@@ -88,15 +73,9 @@ class VoltageInitParametersTest {
     private static final String FILTER_2 = "FILTER_2";
     private static final UUID REPORT_UUID = UUID.fromString("00000000-0000-0000-0000-000000000000");
 
-    private static final VoltageLimitEntity VOLTAGE_LIMIT = new VoltageLimitEntity(UUID.randomUUID(), 5., 10., 0, VoltageLimitParameterType.DEFAULT, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
-    private static final VoltageLimitEntity VOLTAGE_LIMIT_2 = new VoltageLimitEntity(UUID.randomUUID(), 44., 88., 1, VoltageLimitParameterType.DEFAULT, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_2, FILTER_2)));
-    private static final VoltageLimitEntity VOLTAGE_LIMIT_3 = new VoltageLimitEntity(UUID.randomUUID(), -1., -2., 0, VoltageLimitParameterType.MODIFICATION, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
-    private static final VoltageLimitEntity VOLTAGE_LIMIT_4 = new VoltageLimitEntity(UUID.randomUUID(), -20.0, 10.0, 0, VoltageLimitParameterType.MODIFICATION, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
-    private static final VoltageLimitEntity VOLTAGE_LIMIT_5 = new VoltageLimitEntity(UUID.randomUUID(), 10.0, 10.0, 0, VoltageLimitParameterType.DEFAULT, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
-
     private static final JSONComparator REPORTER_COMPARATOR = new CustomComparator(JSONCompareMode.STRICT,
-            // ignore field having uuid changing each run
-            new Customization("reportTree.subReporters[*].taskValues.parameters_id.value", (o1, o2) -> (o1 == null) == (o2 == null))
+        // ignore field having uuid changing each run
+        new Customization("reportTree.subReporters[*].taskValues.parameters_id.value", (o1, o2) -> (o1 == null) == (o2 == null))
     );
 
     private Network network;
@@ -105,13 +84,10 @@ class VoltageInitParametersTest {
     private VoltageInitParametersService voltageInitParametersService;
 
     @Autowired
-    private MockMvc mockMvc;
+    private TestEntityManager entityManager;
 
     @Autowired
     private ObjectMapper mapper;
-
-    @Autowired
-    private VoltageInitParametersRepository parametersRepository;
 
     @MockBean
     private NetworkStoreService networkStoreService;
@@ -121,8 +97,6 @@ class VoltageInitParametersTest {
 
     @BeforeEach
     public void setup() {
-        mapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
-
         network = EurostagTutorialExample1Factory.create(new NetworkFactoryImpl());
         network.getVariantManager().cloneVariant(VariantManagerConstants.INITIAL_VARIANT_ID, VARIANT_ID_1);
         network.getVariantManager().setWorkingVariant(VARIANT_ID_1);
@@ -145,161 +119,6 @@ class VoltageInitParametersTest {
         ));
     }
 
-    @Test
-    void testCreate() throws Exception {
-        VoltageInitParametersInfos parametersToCreate = buildParameters();
-        String parametersToCreateJson = mapper.writeValueAsString(parametersToCreate);
-
-        mockMvc.perform(post(URI_PARAMETERS_BASE).content(parametersToCreateJson).contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk()).andReturn();
-
-        VoltageInitParametersInfos createdParameters = parametersRepository.findAll().get(0).toVoltageInitParametersInfos();
-
-        assertThat(createdParameters).recursivelyEquals(parametersToCreate);
-    }
-
-    @Test
-    void testRead() throws Exception {
-        VoltageInitParametersInfos parametersToRead = buildParameters();
-
-        UUID parametersUuid = saveAndReturnId(parametersToRead);
-
-        MvcResult mvcResult = mockMvc.perform(get(URI_PARAMETERS_GET_PUT + parametersUuid))
-                .andExpect(status().isOk()).andReturn();
-        String resultAsString = mvcResult.getResponse().getContentAsString();
-        VoltageInitParametersInfos receivedParameters = mapper.readValue(resultAsString, new TypeReference<>() {
-        });
-
-        assertThat(receivedParameters).recursivelyEquals(parametersToRead);
-    }
-
-    @Test
-    void testUpdate() throws Exception {
-        VoltageInitParametersInfos parametersToUpdate = buildParameters();
-
-        UUID parametersUuid = saveAndReturnId(parametersToUpdate);
-
-        parametersToUpdate = buildParametersUpdate();
-
-        String parametersToUpdateJson = mapper.writeValueAsString(parametersToUpdate);
-
-        mockMvc.perform(put(URI_PARAMETERS_GET_PUT + parametersUuid).content(parametersToUpdateJson).contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk());
-
-        VoltageInitParametersInfos updatedParameters = parametersRepository.findById(parametersUuid).get().toVoltageInitParametersInfos();
-
-        assertThat(updatedParameters).recursivelyEquals(parametersToUpdate);
-    }
-
-    @Test
-    void testDelete() throws Exception {
-        VoltageInitParametersInfos parametersToDelete = buildParameters();
-
-        UUID parametersUuid = saveAndReturnId(parametersToDelete);
-
-        mockMvc.perform(delete(URI_PARAMETERS_GET_PUT + parametersUuid)).andExpect(status().isOk()).andReturn();
-        assertThat(parametersRepository.count()).as("parameters repository items count").isZero();
-    }
-
-    @Test
-    void testGetAll() throws Exception {
-        VoltageInitParametersInfos parameters1 = buildParameters();
-
-        VoltageInitParametersInfos parameters2 = buildParametersUpdate();
-
-        saveAndReturnId(parameters1);
-
-        saveAndReturnId(parameters2);
-
-        MvcResult mvcResult = mockMvc.perform(get(URI_PARAMETERS_BASE))
-                .andExpect(status().isOk()).andReturn();
-        String resultAsString = mvcResult.getResponse().getContentAsString();
-        List<VoltageInitParametersInfos> receivedParameters = mapper.readValue(resultAsString, new TypeReference<>() {
-        });
-
-        assertThat(receivedParameters).hasSize(2);
-    }
-
-    @Test
-    void testDuplicate() throws Exception {
-        VoltageInitParametersInfos parametersToCreate = buildParameters();
-        String parametersToCreateJson = mapper.writeValueAsString(parametersToCreate);
-        mockMvc.perform(post(URI_PARAMETERS_BASE).content(parametersToCreateJson).contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk()).andReturn();
-        VoltageInitParametersInfos createdParameters = parametersRepository.findAll().get(0).toVoltageInitParametersInfos();
-
-        mockMvc.perform(post(URI_PARAMETERS_BASE)
-                        .param("duplicateFrom", UUID.randomUUID().toString()))
-                .andExpect(status().isNotFound());
-
-        mockMvc.perform(post(URI_PARAMETERS_BASE)
-                        .param("duplicateFrom", createdParameters.getUuid().toString()))
-                .andExpect(status().isOk());
-
-        VoltageInitParametersInfos duplicatedParameters = parametersRepository.findAll().get(1).toVoltageInitParametersInfos();
-        assertThat(duplicatedParameters).recursivelyEquals(createdParameters);
-    }
-
-    /** Save parameters into the repository and return its UUID. */
-    private UUID saveAndReturnId(VoltageInitParametersInfos parametersInfos) {
-        return parametersRepository.save(parametersInfos.toEntity()).getId();
-    }
-
-    private static VoltageInitParametersInfos buildParameters() {
-        return VoltageInitParametersInfos.builder()
-            .voltageLimitsDefault(List.of())
-            .voltageLimitsModification(List.of())
-            .constantQGenerators(List.of(FilterEquipments.builder()
-                    .filterId(UUID.randomUUID())
-                    .filterName("qgenFilter1")
-                    .build(), FilterEquipments.builder()
-                    .filterId(UUID.randomUUID())
-                    .filterName("qgenFilter2")
-                    .build()))
-            .variableTwoWindingsTransformers(List.of(FilterEquipments.builder()
-                    .filterId(UUID.randomUUID())
-                    .filterName("vtwFilter1")
-                    .build(), FilterEquipments.builder()
-                    .filterId(UUID.randomUUID())
-                    .filterName("vtwFilter2")
-                    .build()))
-            .build();
-    }
-
-    private static VoltageInitParametersInfos buildParametersUpdate() {
-        return VoltageInitParametersInfos.builder()
-            .voltageLimitsModification(List.of(VoltageLimitInfos.builder()
-                .priority(0)
-                .lowVoltageLimit(2.0)
-                .highVoltageLimit(20.0)
-                .filters(List.of(FilterEquipments.builder()
-                    .filterId(UUID.randomUUID())
-                    .filterName("filterName")
-                    .build()))
-                .build()))
-            .voltageLimitsDefault(List.of(VoltageLimitInfos.builder()
-                .priority(0)
-                .lowVoltageLimit(2.0)
-                .highVoltageLimit(20.0)
-                .filters(List.of(FilterEquipments.builder()
-                    .filterId(UUID.randomUUID())
-                    .filterName("filterName")
-                    .build()))
-                .build()))
-            .variableShuntCompensators(List.of(FilterEquipments.builder()
-                .filterId(UUID.randomUUID())
-                .filterName("vscFilter1")
-                .build()))
-            .variableTwoWindingsTransformers(List.of(FilterEquipments.builder()
-                    .filterId(UUID.randomUUID())
-                    .filterName("vtwFilter1Modified")
-                    .build(), FilterEquipments.builder()
-                    .filterId(UUID.randomUUID())
-                    .filterName("vtwFilter2Modified")
-                    .build()))
-            .build();
-    }
-
     private static Consumer<VoltageLimitOverride> assertVoltageLimitOverride(final String levelId, final VoltageLimitType limitType) {
         return assertVoltageLimitOverride(levelId, limitType, null);
     }
@@ -320,8 +139,8 @@ class VoltageInitParametersTest {
     }
 
     private ListAssert<VoltageLimitOverride> testsBuildSpecificVoltageLimitsCommon(List<VoltageLimitEntity> voltageLimits, String reportFilename) throws Exception {
-        final VoltageInitParametersEntity voltageInitParameters = parametersRepository.save(
-                new VoltageInitParametersEntity(UUID.randomUUID(), null, "", voltageLimits, null, null, null)
+        final VoltageInitParametersEntity voltageInitParameters = entityManager.persistFlushFind(
+            new VoltageInitParametersEntity(null, null, "", voltageLimits, null, null, null)
         );
         final VoltageInitRunContext context = new VoltageInitRunContext(NETWORK_UUID, VARIANT_ID_1, null, REPORT_UUID, null, "", "", voltageInitParameters.getId());
         final OpenReacParameters openReacParameters = voltageInitParametersService.buildOpenReacParameters(context, network);
@@ -333,7 +152,9 @@ class VoltageInitParametersTest {
     @DisplayName("buildSpecificVoltageLimits: No voltage limit modification")
     @Test
     void testsBuildSpecificVoltageLimitsSimple() throws Exception {
-        testsBuildSpecificVoltageLimitsCommon(List.of(VOLTAGE_LIMIT, VOLTAGE_LIMIT_2), "reporter_buildOpenReacParameters.json")
+        final VoltageLimitEntity voltageLimit = new VoltageLimitEntity(null, 5., 10., 0, VoltageLimitParameterType.DEFAULT, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
+        final VoltageLimitEntity voltageLimit2 = new VoltageLimitEntity(null, 44., 88., 1, VoltageLimitParameterType.DEFAULT, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_2, FILTER_2)));
+        testsBuildSpecificVoltageLimitsCommon(List.of(voltageLimit, voltageLimit2), "reporter_buildOpenReacParameters.json")
             .hasSize(4)
             //No override should be relative since there is no voltage limit modification
             .noneMatch(VoltageLimitOverride::isRelative)
@@ -350,8 +171,11 @@ class VoltageInitParametersTest {
     @DisplayName("buildSpecificVoltageLimits: With voltage limit modifications")
     @Test
     void testsBuildSpecificVoltageLimitsWithLimitModifications() throws Exception {
+        final VoltageLimitEntity voltageLimit = new VoltageLimitEntity(null, 5., 10., 0, VoltageLimitParameterType.DEFAULT, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
+        final VoltageLimitEntity voltageLimit2 = new VoltageLimitEntity(null, 44., 88., 1, VoltageLimitParameterType.DEFAULT, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_2, FILTER_2)));
+        final VoltageLimitEntity voltageLimit3 = new VoltageLimitEntity(null, -1., -2., 0, VoltageLimitParameterType.MODIFICATION, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
         //We now add limit modifications in additions to defaults settings
-        testsBuildSpecificVoltageLimitsCommon(List.of(VOLTAGE_LIMIT, VOLTAGE_LIMIT_2, VOLTAGE_LIMIT_3), "reporter_buildOpenReacParameters_withLimitModifications.json")
+        testsBuildSpecificVoltageLimitsCommon(List.of(voltageLimit, voltageLimit2, voltageLimit3), "reporter_buildOpenReacParameters_withLimitModifications.json")
             //Limits that weren't impacted by default settings are now impacted by modification settings
             .hasSize(8)
             //There should (not?) be relative overrides since voltage limit modification are applied
@@ -367,8 +191,9 @@ class VoltageInitParametersTest {
     @DisplayName("buildSpecificVoltageLimits: Case relative true overrides")
     @Test
     void testsBuildSpecificVoltageLimitsCaseRelativeTrue() throws Exception {
+        final VoltageLimitEntity voltageLimit4 = new VoltageLimitEntity(null, -20.0, 10.0, 0, VoltageLimitParameterType.MODIFICATION, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
         // We need to check for the case of relative = true with the modification less than 0 => the new low voltage limit = low voltage limit * -1
-        testsBuildSpecificVoltageLimitsCommon(List.of(VOLTAGE_LIMIT_4), "reporter_buildOpenReacParameters_caseRelativeTrue.json")
+        testsBuildSpecificVoltageLimitsCommon(List.of(voltageLimit4), "reporter_buildOpenReacParameters_caseRelativeTrue.json")
             .hasSize(4)
             // isRelative: There should have relative true overrides since voltage limit modification are applied for VLGEN
             // getLimit: The low voltage limit must be impacted by the modification of the value
@@ -376,11 +201,13 @@ class VoltageInitParametersTest {
         //note: VoltageLimitOverride implement equals() correctly, so we can use it
     }
 
-    @DisplayName("buildSpecificVoltageLimits: No voltage limit modification")
+    @DisplayName("buildSpecificVoltageLimits: Case relative false overrides")
     @Test
     void testsBuildSpecificVoltageLimitsCaseRelativeFalse() throws Exception {
+        final VoltageLimitEntity voltageLimit4 = new VoltageLimitEntity(null, -20.0, 10.0, 0, VoltageLimitParameterType.MODIFICATION, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
+        final VoltageLimitEntity voltageLimit5 = new VoltageLimitEntity(null, 10.0, 10.0, 0, VoltageLimitParameterType.DEFAULT, List.of(new FilterEquipmentsEmbeddable(FILTER_UUID_1, FILTER_1)));
         // We need to check for the case of relative = false with the modification less than 0 => the new low voltage limit = 0
-        testsBuildSpecificVoltageLimitsCommon(List.of(VOLTAGE_LIMIT_4, VOLTAGE_LIMIT_5), "reporter_buildOpenReacParameters_caseRelativeFalse.json")
+        testsBuildSpecificVoltageLimitsCommon(List.of(voltageLimit4, voltageLimit5), "reporter_buildOpenReacParameters_caseRelativeFalse.json")
             .hasSize(8)
             // isRelative: There should have relative false overrides since voltage limit modification are applied for VLHV1
             // getLimit: The low voltage limit must be impacted by the modification of the value
